requirements.txt
ccxt
python-dotenv
numpy
pandas
